+++ CHANGELOG.md
@@ -1,11 +1,17 @@
 # Changelog
 
+## [Unreleased]
+
+### Added
+
+- [TD-4739] Support validation of potentially unsafe content
+
 ## [4.45.0] 2022-05-31
 
 ### Changed
 
 - [TD-3920] `TdDfLib.Format.apply_template/3` takes only template fields of the
-            content.
+  content
 
 ## [4.44.0] 2022-05-20